refactor: domain module 추출
