+++ kitchenpos-domain/build.gradle
@@ -1,0 +1,14 @@
+
+plugins {
+    id 'org.springframework.boot'
+    id 'java'
+}
+
+bootJar { enabled = false }
+jar { enabled = true }
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    runtimeOnly 'com.h2database:h2'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+}
