+++ kitchenpos-domain/src/main/java/kitchenpos/domain/model/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
