+++ kitchenpos-domain/src/main/java/kitchenpos/domain/model/Money.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.model;
 
 import java.math.BigDecimal;
 import java.util.Objects;
