+++ kitchenpos-domain/src/main/java/kitchenpos/domain/model/OrderLineItem.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.model;
 
 import java.util.Objects;
 import javax.persistence.Column;
