+++ kitchenpos-domain/src/main/java/kitchenpos/domain/model/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.model;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION
