+++ kitchenpos-domain/src/main/java/kitchenpos/domain/model/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.model;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
