+++ kitchenpos-domain/src/test/java/kitchenpos/domain/DomainApplicationTests.java
@@ -1,0 +1,13 @@
+package kitchenpos.domain;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DomainApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
