+++ src/main/java/kitchenpos/dto/MenuCreateRequest.java
@@ -3,9 +3,9 @@ package kitchenpos.dto;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Money;
+import kitchenpos.domain.model.Menu;
+import kitchenpos.domain.model.MenuProduct;
+import kitchenpos.domain.model.Money;
 
 public class MenuCreateRequest {
