+++ src/main/java/kitchenpos/dto/MenuGroupCreateRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.model.MenuGroup;
 
 /**
  * @author : leesangbae
