+++ src/main/java/kitchenpos/dto/OrderCreateRequest.java
@@ -4,9 +4,9 @@ import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.model.Order;
+import kitchenpos.domain.model.OrderLineItem;
+import kitchenpos.domain.model.OrderStatus;
 
 public class OrderCreateRequest {
