+++ src/main/java/kitchenpos/dto/OrderLineItemCreateRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.model.OrderLineItem;
 
 public class OrderLineItemCreateRequest {
