+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -9,13 +9,13 @@ import kitchenpos.application.creator.MenuGroupHelper;
 import kitchenpos.application.creator.MenuHelper;
 import kitchenpos.application.creator.MenuProductHelper;
 import kitchenpos.application.creator.ProductHelper;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.model.MenuGroup;
+import kitchenpos.domain.model.Product;
 import kitchenpos.dto.MenuCreateRequest;
 import kitchenpos.dto.MenuDto;
 import kitchenpos.dto.MenuProductRequest;
-import kitchenpos.repository.MenuGroupDao;
-import kitchenpos.repository.ProductDao;
+import kitchenpos.domain.repository.MenuGroupDao;
+import kitchenpos.domain.repository.ProductDao;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
