+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -12,19 +12,19 @@ import kitchenpos.application.creator.MenuProductHelper;
 import kitchenpos.application.creator.OrderLineItemHelper;
 import kitchenpos.application.creator.OrderTableHelper;
 import kitchenpos.application.creator.ProductHelper;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.model.MenuGroup;
+import kitchenpos.domain.model.OrderTable;
+import kitchenpos.domain.model.Product;
 import kitchenpos.dto.MenuCreateRequest;
 import kitchenpos.dto.MenuDto;
 import kitchenpos.dto.MenuProductRequest;
 import kitchenpos.dto.OrderCreateRequest;
 import kitchenpos.dto.OrderDto;
 import kitchenpos.dto.OrderLineItemCreateRequest;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.repository.MenuGroupDao;
-import kitchenpos.repository.OrderTableDao;
-import kitchenpos.repository.ProductDao;
+import kitchenpos.domain.model.OrderStatus;
+import kitchenpos.domain.repository.MenuGroupDao;
+import kitchenpos.domain.repository.OrderTableDao;
+import kitchenpos.domain.repository.ProductDao;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
