+++ src/test/java/kitchenpos/application/creator/OrderHelper.java
@@ -3,7 +3,7 @@ package kitchenpos.application.creator;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.model.OrderTable;
 import kitchenpos.dto.OrderCreateRequest;
 import kitchenpos.dto.OrderDto;
 import kitchenpos.dto.OrderLineItemCreateRequest;
